Add files via upload
The project Spyware Prevention is mainly related to preventing spyware attacks on webcams. The main moto of the project is to provide privacy for the user through the unique 6-digit password generated to the email for authentication. through this step, we ensure the privacy of the user. The project consists of a user-friendly interface for enabling and disabling camera access. for each respective action, it generates the unique password to the registered email in the project flow process.

through this project, the webcam can be locked by the respective user and can prevent access to the webcam for the unauthorized user through the authentication mechanism. 

for any doubt and information for source code. please do contact msainath5488@gmail.com